Spell interaction with cooldown on Barrier + notifications + interface guide
Added spell interaction with the barrier spell, so you can shield yourself showing as a white boarder around the player info bar. Made notifications for all the spells when clicked on the bar with the asigned keyboard key. Also implemented a interface guide, so the user know how to interact iwth the interface